refactor: reorganize the Management namespace
* moved inspectors.php to Assegai\Orm\Management\Inspectors
* moved options to Assegai\Orm\Management\Options
