+++ src/DataSource/Schema.php
@@ -10,10 +10,10 @@ use Assegai\Orm\Exceptions\DataSourceConnectionException;
 use Assegai\Orm\Exceptions\GeneralSQLQueryException;
 use Assegai\Orm\Exceptions\ORMException;
 use Assegai\Orm\Interfaces\IDataObject;
-use Assegai\Orm\Management\ColumnInspector;
-use Assegai\Orm\Management\EntityInspector;
-use Assegai\Orm\Management\EntityManager;
 use Assegai\Orm\Interfaces\ISchema;
+use Assegai\Orm\Management\EntityManager;
+use Assegai\Orm\Management\Inspectors\ColumnInspector;
+use Assegai\Orm\Management\Inspectors\EntityInspector;
 use Assegai\Orm\Metadata\SchemaMetadata;
 use Assegai\Orm\Metadata\SQLTableDescription;
 use Assegai\Orm\Migrations\SchemaChangeManifest;
