+++ src/Interfaces/IRepository.php
@@ -5,15 +5,14 @@ namespace Assegai\Orm\Interfaces;
 use Assegai\Orm\Exceptions\ClassNotFoundException;
 use Assegai\Orm\Exceptions\IllegalTypeException;
 use Assegai\Orm\Exceptions\ORMException;
-use Assegai\Orm\Management\FindManyOptions;
-use Assegai\Orm\Management\FindOneOptions;
-use Assegai\Orm\Management\FindOptions;
-use Assegai\Orm\Management\FindWhereOptions;
-use Assegai\Orm\Management\RemoveOptions;
+use Assegai\Orm\Management\Options\FindManyOptions;
+use Assegai\Orm\Management\Options\FindOneOptions;
+use Assegai\Orm\Management\Options\FindOptions;
+use Assegai\Orm\Management\Options\FindWhereOptions;
+use Assegai\Orm\Management\Options\RemoveOptions;
 use Assegai\Orm\Queries\QueryBuilder\Results\DeleteResult;
 use Assegai\Orm\Queries\QueryBuilder\Results\InsertResult;
 use Assegai\Orm\Queries\QueryBuilder\Results\UpdateResult;
-
 use stdClass as Entity;
 
 interface IRepository
@@ -162,7 +161,7 @@ interface IRepository
   /**
    * Finds entities that match given `FindWhereOptions`.
    *
-   * @param FindWhereOptions|array $where
+   * @param \Assegai\Orm\Management\Options\FindWhereOptions|array $where
    * @return null|array<Entity> Returns a list of entities that match the given `FindWhereOptions`.
    */
   public function findBy(FindWhereOptions|array $where): ?array;
@@ -172,7 +171,7 @@ interface IRepository
    * Also counts all entities that match given conditions,
    * but ignores pagination settings (from and take options).
    *
-   * @param FindManyOptions|array|null $options
+   * @param \Assegai\Orm\Management\Options\FindManyOptions|array|null $options
    * @return array<[Entity,int]>
    */
   public function findAndCount(FindManyOptions|array|null $options = null): array;
@@ -182,7 +181,7 @@ interface IRepository
    * Also counts all entities that match given conditions,
    * but ignores pagination settings (from and take options).
    *
-   * @param FindWhereOptions|array $where
+   * @param \Assegai\Orm\Management\Options\FindWhereOptions|array $where
    * @return array
    */
   public function findAndCountBy(FindWhereOptions|array $where): array;
@@ -191,7 +190,7 @@ interface IRepository
    * Finds first entity by a given find options.
    * If entity was not found in the database - returns null.
    *
-   * @param FindOptions|FindOneOptions|array $options
+   * @param \Assegai\Orm\Management\Options\FindOptions|\Assegai\Orm\Management\Options\FindOneOptions|array $options
    * @return null|Entity Returns the entity if found, null otherwise.
    */
   public function findOne(FindOptions|FindOneOptions|array $options): ?object;
