+++ src/Management/Repository.php
@@ -14,6 +14,11 @@ use Assegai\Orm\Exceptions\ORMException;
 use Assegai\Orm\Exceptions\SaveException;
 use Assegai\Orm\Interfaces\IFactory;
 use Assegai\Orm\Interfaces\IRepository;
+use Assegai\Orm\Management\Options\FindManyOptions;
+use Assegai\Orm\Management\Options\FindOneOptions;
+use Assegai\Orm\Management\Options\FindOptions;
+use Assegai\Orm\Management\Options\FindWhereOptions;
+use Assegai\Orm\Management\Options\RemoveOptions;
 use Assegai\Orm\Queries\QueryBuilder\Results\DeleteResult;
 use Assegai\Orm\Queries\QueryBuilder\Results\InsertResult;
 use Assegai\Orm\Queries\QueryBuilder\Results\UpdateResult;
